requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
cloudinary==1.36.0
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
aiosmtplib==3.0.1
email-validator==2.1.0
sendgrid==6.11.0
pandas==2.1.4
openpyxl==3.1.2
pyproj==3.6.1
geoalchemy2==0.14.2
geopandas==0.14.1
shapely==2.0.2
fiona==1.9.5
